requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pdfplumber
python-docx
requests
pydantic
